Precursor base class. filtered and unfiltered are simplified
